NA-6996: Fix python path issue
